Move modules out of node_modules
