+++ src/editor/index.js
